rename tests dir to fix coverage
