+++ lib/reporters/_baseReport.js
