+++ lib/reporters/badge.js
