+++ lib/testers/_baseTest.js
