+++ lib/testers/coverageExcludes.js
