+++ lib/testers/jshint.js
